requirements.txt
psycopg[binary]>=3.1
sentence-transformers>=3.0
torch
tqdm>=4.66
python-dotenv>=1.0
google-genai>=1.0
